Genarated dbus objects for hamster, afresh
To generate:
apt-get install dbus-java-bin
CreateInterface --session org.gnome.Hamster /org/gnome/Hamster

I matters for what version of hamster you generate those classes.
I generated for hamster from my fork repository:
https://github.com/rynkowsg/hamster.

+++ data/src/main/java/org/gnome/Hamster.java
@@ -1,40 +1,42 @@
+/* File: org/gnome/Hamster.java */
 package org.gnome;
 import java.util.List;
 import java.util.Map;
 import org.freedesktop.dbus.DBusInterface;
 import org.freedesktop.dbus.DBusSignal;
+import org.freedesktop.dbus.UInt32;
 import org.freedesktop.dbus.Variant;
 import org.freedesktop.dbus.exceptions.DBusException;
 public interface Hamster extends DBusInterface
 {
-   public static class FactsChanged extends DBusSignal
-   {
-      public FactsChanged(String path) throws DBusException
-      {
-         super(path);
-      }
-   }
-   public static class ToggleCalled extends DBusSignal
-   {
-      public ToggleCalled(String path) throws DBusException
-      {
-         super(path);
-      }
-   }
-   public static class ActivitiesChanged extends DBusSignal
-   {
-      public ActivitiesChanged(String path) throws DBusException
-      {
-         super(path);
-      }
-   }
-   public static class TagsChanged extends DBusSignal
-   {
-      public TagsChanged(String path) throws DBusException
-      {
-         super(path);
-      }
-   }
+  public static class FactsChanged extends DBusSignal
+  {
+    public FactsChanged(String path) throws DBusException
+    {
+      super(path);
+    }
+  }
+  public static class ToggleCalled extends DBusSignal
+  {
+    public ToggleCalled(String path) throws DBusException
+    {
+      super(path);
+    }
+  }
+  public static class ActivitiesChanged extends DBusSignal
+  {
+    public ActivitiesChanged(String path) throws DBusException
+    {
+      super(path);
+    }
+  }
+  public static class TagsChanged extends DBusSignal
+  {
+    public TagsChanged(String path) throws DBusException
+    {
+      super(path);
+    }
+  }
 
   public List<Struct1> GetTags(boolean only_autocomplete);
   public void Quit();
@@ -52,7 +54,7 @@ public interface Hamster extends DBusInterface
   public int UpdateFact(int fact_id, String fact, int start_time, int end_time, boolean temporary, boolean exported);
   public void RemoveFact(int fact_id);
   public void UpdateActivity(int id, String name, int category_id);
-  public List<Struct7> GetFacts(long start_date, long end_date, String search_terms, long limit, boolean asc_by_date);
+  public List<Struct7> GetFacts(UInt32 start_date, UInt32 end_date, String search_terms, UInt32 limit, boolean asc_by_date);
   public int AddCategory(String name);
   public void RemoveActivity(int id);
   public void RemoveCategory(int id);
